requirements.txt
flask==3.1.1
groq==0.31.0
python-dotenv==1.1.1
gunicorn==23.0.0
requests==2.32.4